[fix] Simplify unit tests by removing uncheckpointed client
